+++ CHANGELOG.md
@@ -4,12 +4,13 @@
 
 ## Unreleased
 
+## 1.18.0 - 2023-06-12
+
 ### Added
 - Add setting for AFX attribute value completion
 
 ### Fixed
 - AFX tag completion and inspection suppression
-
 
 ## 1.17.0 - 2023-06-07
 - Allow quotes in AFX attribute names